Refactored to behave closer to MVC model - Added ViewInterface.java - Added GUI.java, implements ViewInterface - CommandLine.java now implements ViewInterface - View no longer contains Controller
