+++ src/main/java/mike/gui/ViewInterface.java
@@ -1,0 +1,5 @@
+package mike.gui;
+
+public interface ViewInterface {
+	
+}
