+++ src/main/java/mike/main.java
@@ -1,15 +1,15 @@
 package mike;
 
-import mike.cli.CommandLine;
+import mike.gui.Controller;
 import mike.gui.View;
 
 public class main {
-
+	static Controller controller;
 	public static void main(String[] args) {
 		if (args.length == 1 && args[0].equals("cli")) {
-			CommandLine.commandInterface();
+			controller = new Controller(View.InterfaceType.CLI);
 		} else if (args.length == 0) {
-			View.guiInterface();
+			controller = new Controller(View.InterfaceType.GUI);
 		} else {
 			System.out.println(
 					"Invalid input. Enter 'cli' for the command line interface, or enter nothing for the gui.");
